+++ app/controllers/files.js
@@ -21,7 +21,7 @@ module.exports = function() {
     core.on('files:new', function(file, room, user) {
         var fil = file.toJSON();
         fil.owner = user;
-        fil.room = room;
+        fil.room = room.toJSON(user);
 
         app.io.to(room._id)
               .emit('files:new', fil);
@@ -105,6 +105,9 @@ module.exports = function() {
         },
         list: function(req, res) {
             var options = {
+                    userId: req.user._id,
+                    password: req.param('password'),
+
                     room: req.param('room'),
                     reverse: req.param('reverse'),
                     skip: req.param('skip'),
@@ -116,6 +119,11 @@ module.exports = function() {
                 if (err) {
                     return res.sendStatus(400);
                 }
+
+                files = files.map(function(file) {
+                    return file.toJSON(req.user);
+                });
+
                 res.json(files);
             });
         }