+++ app/core/presence.js
@@ -1,7 +1,6 @@
 'use strict';
 
-var _ = require('lodash'),
-    Connection = require('./presence/connection'),
+var Connection = require('./presence/connection'),
     Room = require('./presence/room'),
     ConnectionCollection = require('./presence/connection-collection'),
     RoomCollection = require('./presence/room-collection'),
@@ -12,7 +11,7 @@ function PresenceManager(options) {
     this.system = new Room({ system: true });
     this.connections = new ConnectionCollection();
     this.rooms = new RoomCollection();
-    this.users = new UserCollection();
+    this.users = new UserCollection({ core: this.core });
     this.rooms.on('user_join', this.onJoin.bind(this));
     this.rooms.on('user_leave', this.onLeave.bind(this));
 
@@ -52,9 +51,9 @@ PresenceManager.prototype.disconnect = function(connection) {
     this.rooms.removeConnection(connection);
 };
 
-PresenceManager.prototype.join = function(connection, roomId, roomSlug) {
-    var room = this.rooms.getOrAdd(roomId, roomSlug);
-    room.addConnection(connection);
+PresenceManager.prototype.join = function(connection, room) {
+    var pRoom = this.rooms.getOrAdd(room);
+    pRoom.addConnection(connection);
 };
 
 PresenceManager.prototype.leave = function(connection, roomId) {