+++ app/core/presence/connection-collection.js
@@ -41,6 +41,26 @@ ConnectionCollection.prototype.getUsers = function(filter) {
     return users;
 };
 
+ConnectionCollection.prototype.getUsers = function(filter) {
+    var connections = this.connections;
+
+    if (filter) {
+        connections = this.query(filter);
+    }
+
+    var users = _.chain(connections)
+                .filter(function(value) {
+                    return !!value.user;
+                })
+                .map(function(value) {
+                    return value.user;
+                })
+                .uniq('id')
+                .value();
+
+    return users;
+};
+
 ConnectionCollection.prototype.getUserIds = function(filter) {
     var users = this.getUsers(filter);