+++ app/models/room.js
@@ -5,7 +5,8 @@
 'use strict';
 
 var mongoose = require('mongoose'),
-    uniqueValidator = require('mongoose-unique-validator');
+    uniqueValidator = require('mongoose-unique-validator'),
+    bcrypt = require('bcryptjs');
 
 var ObjectId = mongoose.Schema.Types.ObjectId;
 
@@ -36,6 +37,10 @@ var RoomSchema = new mongoose.Schema({
 		ref: 'User',
         required: true
     },
+    participants: [{ // We can have an array per role
+		type: ObjectId,
+		ref: 'User'
+	}],
 	messages: [{
 		type: ObjectId,
 		ref: 'Message'
@@ -47,6 +52,14 @@ var RoomSchema = new mongoose.Schema({
     lastActive: {
         type: Date,
         default: Date.now
+    },
+    private: {
+        type: Boolean,
+        default: false
+    },
+    password: {
+        type: String,
+        required: false//only for password-protected room
     }
 });
 
@@ -54,21 +67,141 @@ RoomSchema.virtual('handle').get(function() {
     return this.slug || this.name.replace(/\W/i, '');
 });
 
+RoomSchema.virtual('hasPassword').get(function() {
+    return !!this.password;
+});
+
+RoomSchema.pre('save', function(next) {
+    var room = this;
+    if (!room.password || !room.isModified('password')) {
+        return next();
+    }
+
+    bcrypt.hash(room.password, 10, function(err, hash) {
+        if (err) {
+            return next(err);
+        }
+        room.password = hash;
+        next();
+    });
+});
+
 RoomSchema.plugin(uniqueValidator, {
     message: 'Expected {PATH} to be unique'
 });
 
-RoomSchema.method('toJSON', function() {
+RoomSchema.method('isAuthorized', function(userId) {
+    if (!userId) {
+        return false;
+    }
+
+    userId = userId.toString();
+
+    // Check if userId doesn't match MongoID format
+    if (!/^[a-f\d]{24}$/i.test(userId)) {
+        return false;
+    }
+
+    if (!this.password && !this.private) {
+        return true;
+    }
+
+    if (this.owner.equals(userId)) {
+        return true;
+    }
+
+    return this.participants.some(function(participant) {
+        if (participant._id) {
+            return participant._id.equals(userId);
+        }
+
+        if (participant.equals) {
+            return participant.equals(userId);
+        }
+
+        if (participant.id) {
+            return participant.id === userId;
+        }
+
+        return participant === userId;
+    });
+});
+
+RoomSchema.method('canJoin', function(options, cb) {
+    var userId = options.userId,
+        password = options.password,
+        saveMembership = options.saveMembership;
+
+    if (this.isAuthorized(userId)) {
+        return cb(null, true);
+    }
+
+    if (!this.password) {
+        return cb(null, false);
+    }
+
+    bcrypt.compare(password || '', this.password, function(err, isMatch) {
+        if(err) {
+            return cb(err);
+        }
+
+        if (!isMatch) {
+            return cb(null, false);
+        }
+
+        if (!saveMembership) {
+            return cb(null, true);
+        }
+
+        this.participants.push(userId);
+
+        this.save(function(err) {
+            if(err) {
+                return cb(err);
+            }
+
+            cb(null, true);
+        });
+
+    }.bind(this));
+});
+
+RoomSchema.method('toJSON', function(user) {
+    var userId = user ? (user._id || user.id || user) : null;
+    var authorized = false;
+
+    if (userId) {
+        authorized = this.isAuthorized(userId);
+    }
+
     var room = this.toObject();
-    return {
+
+    var data = {
         id: room._id,
         slug: room.slug,
         name: room.name,
         description: room.description,
         lastActive: room.lastActive,
         created: room.created,
-        owner: room.owner
+        owner: room.owner,
+        private: room.private,
+        hasPassword: this.hasPassword,
+        participants: []
     };
+
+    if (room.private && authorized) {
+        var participants = this.participants || [];
+        data.participants = participants.map(function(user) {
+            return user.username ? user.username : user;
+        });
+    }
+
+    if (this.users) {
+        data.users = this.users;
+        data.userCount = this.users.length;
+    }
+
+    return data;
  });
 
 RoomSchema.statics.findByIdOrSlug = function(identifier, cb) {