+++ app/xmpp/events/user-avatar-ready.js
@@ -3,12 +3,11 @@
 var _ = require('lodash'),
     Stanza = require('node-xmpp-core').Stanza,
     settings = require('./../../config'),
-    helper = require('./../helper'),
     EventListener = require('./../event-listener');
 
 module.exports = EventListener.extend({
 
-    on: 'xmpp:avatar_ready',
+    on: 'avatar-cache:update',
 
     then: function(user) {
         if (!settings.private.enable) {
@@ -29,20 +28,20 @@ module.exports = EventListener.extend({
             type: 'xmpp'
         });
 
-        _.each(connections, function(x) {
-            if (x.user.id === user.id) {
+        _.each(connections, function(connection) {
+            if (connection.user.id === user.id) {
                 return;
             }
 
             // Reannounce presence
             var presence = new Stanza.Presence({
-                from: helper.getUserJid(user.username)
+                from: connection.getUserJid(user.username)
             });
 
-            helper.populateVcard(presence, user);
+            connection.populateVcard(presence, user, this.core);
 
-            this.send(x, presence);
+            this.send(connection, presence);
         }, this);
     }
 
-}); 
+});