+++ app/xmpp/msg-processors/roster-get.js
@@ -2,8 +2,7 @@
 
 var _ = require('lodash'),
     MessageProcessor = require('./../msg-processor'),
-    settings = require('./../../config'),
-    helper = require('./../helper');
+    settings = require('./../../config');
 
 module.exports = MessageProcessor.extend({
 
@@ -32,7 +31,6 @@ module.exports = MessageProcessor.extend({
     },
 
     sendAllUsers: function(cb) {
-        var users = this.core.users.list({}, function(err, users) {
             if (err) {
                 return cb(err);
             }
@@ -57,7 +55,7 @@ module.exports = MessageProcessor.extend({
             }
 
             v.c('item', {
-                jid: helper.getUserJid(user.username),
+                jid: this.connection.getUserJid(user.username),
                 name: user.displayName,
                 subscription: 'both'
             }).c('group').t('Let\'s Chat');